WIP - test: tweaking git-races test to maybe work on appveyor
This really shouldn't be going to the network like this
but I would need to take more time to carefully reproduce
this with local repositories.

It's really icky that we're relying on an external git repo
owned by someone else to pull this off.

The actual fix in this patch is just to bypass ssh key errors
that happen in AppVeyor. I'm not sure why Travis doesn't have
this issue, but that's probably because Travis does its own
ssh key business?